BIGINT is toxic!

+++ ORM/Auths.js
@@ -6,14 +6,14 @@ var sequelize = require(path.resolve(__dirname, '../libs/sequelize'));
 
 /**
  * Define Auths schema
- * @param  {Sequelize.BIGINT} aid Auth's ID
+ * @param  {Sequelize.INTEGER} aid Auth's ID
  * @param  {Sequelize.STRING} user_identity User's Identity
  * @param  {Sequelize.TEXT} salt Random generated salt
  * @param  {Sequelize.TEXT} answer Real password + salt = answer
  */
 var Auths = sequelize.define('auths', {
 	aid: {
-		type: Sequelize.BIGINT,
+		type: Sequelize.INTEGER,
 		allowNull: false,
 		unique: true,
 		autoIncrement: true,
